requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
torchvision>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Optional: For advanced features
# jupyter>=1.0.0                  # Jupyter notebooks
# tensorboard>=2.13.0             # TensorBoard logging
# wandb>=0.15.0                   # Weights & Biases logging
# scikit-learn>=1.3.0             # Additional metrics
# scipy>=1.10.0                   # Scientific computing
